scripts/requirements.txt
Flask
markdown
google-cloud-datastore
google-cloud-pubsub
pandas
email-validator
xlrd
openpyxl
